DRY for hash algorithms, more tests for hash
Created a macro that reduces code repetition for Sha2 algorithms. Also added the .block_size_bytes() method that returns the Blcok Size of the algorithm in bytes. Added tests for this method and .h_len().

+++ backend/src/cryptography/oaep/mgf.rs
@@ -7,7 +7,7 @@ pub fn mgf1(
     seed: &[u8],
     length: usize,
 ) -> Vec<u8> {
-    let h_len = hash.h_len();
+    let h_len = hash.h_len() as usize;
 
     assert!((length >> 32) <= h_len, "mask too long");
